3. Project Executable files/requirements.txt
streamlit
google-generativeai
langdetect
googletrans==4.0.0-rc1
python-dotenv